:truck: Refactor user routes and improve error handling
- Move user routes from controllers to routes directory
- Update import path for usersRoutes in app.ts
- Remove console.log in production for ZodError
- Remove unnecessary console.error in test environment
- Adjust error handling logic in app.ts
