+++ src/app.ts
@@ -1,4 +1,3 @@
-import { usersRoutes } from '@/http/controllers/users/routes'
 import { serviceAdsRoutes } from '@/http/routes/serviceAds'
 import cors from '@fastify/cors'
 import fastifyJwt from '@fastify/jwt'
@@ -7,6 +6,7 @@ import { ZodError } from 'zod'
 import { env } from './env/config'
 import { serviceSubTypesRoutes } from './http/routes/serviceSubTypes'
 import { serviceTypesRoutes } from './http/routes/serviceTypes'
+import { usersRoutes } from './http/routes/users'
 
 export const app = fastify()
 
@@ -35,7 +35,8 @@ app.register(serviceSubTypesRoutes)
 
 app.setErrorHandler((error, _, reply) => {
   if (error instanceof ZodError) {
-    console.log(JSON.stringify(error, null, 2))
+    if (env.NODE_ENV !== 'test') console.log(JSON.stringify(error, null, 2))
+
     return reply
       .status(500)
       .send({ message: 'Validation error', issues: error.format() })
@@ -43,7 +44,6 @@ app.setErrorHandler((error, _, reply) => {
 
   if (env.NODE_ENV !== 'production') {
     console.error(error)
-    console.error('não imprimir no teste')
   } else {
     // TODO: log unknown error
   }
